refactor: 컨플릭트 수정
