+++ src/test/java/coffeemeet/server/certification/implement/CertificationCommandTest.java
@@ -1,17 +1,11 @@
 package coffeemeet.server.certification.implement;
 
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.businessCardUrl;
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.certification;
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.companyEmail;
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.companyName;
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.department;
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.pendingCertification;
-import static coffeemeet.server.common.fixture.entity.UserFixture.user;
 import static coffeemeet.server.common.fixture.CertificationFixture.businessCardUrl;
 import static coffeemeet.server.common.fixture.CertificationFixture.certification;
 import static coffeemeet.server.common.fixture.CertificationFixture.companyEmail;
 import static coffeemeet.server.common.fixture.CertificationFixture.companyName;
 import static coffeemeet.server.common.fixture.CertificationFixture.department;
+import static coffeemeet.server.common.fixture.CertificationFixture.pendingCertification;
 import static coffeemeet.server.common.fixture.UserFixture.user;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
