+++ src/test/java/coffeemeet/server/certification/implement/VerificationInfoQueryTest.java
@@ -1,6 +1,6 @@
 package coffeemeet.server.certification.implement;
 
-import static coffeemeet.server.common.fixture.entity.CertificationFixture.emailVerification;
+import static coffeemeet.server.common.fixture.CertificationFixture.emailVerification;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
