+++ src/test/java/coffeemeet/server/chatting/current/service/ChattingMessageServiceTest.java
@@ -20,8 +20,6 @@ import coffeemeet.server.chatting.current.implement.ChattingRoomQuery;
 import coffeemeet.server.chatting.current.implement.ChattingSessionCommand;
 import coffeemeet.server.chatting.current.implement.ChattingSessionQuery;
 import coffeemeet.server.chatting.current.service.dto.ChattingDto;
-import coffeemeet.server.common.implement.FCMNotificationSender;
-import coffeemeet.server.chatting.current.service.dto.ChattingDto.Response;
 import coffeemeet.server.common.infrastructure.FCMNotificationSender;
 import coffeemeet.server.user.domain.User;
 import coffeemeet.server.user.implement.UserCommand;
